Fix overlay block copying
Overlay file which is a spdk disk, had 8 MB of data and it affected
the offsetting of sector number during enc/dec operations and as a
result, directly copynig those blocks wouldn't work. that's why we
need to first decrypt with the right sector number accounting for
that 8MB of metadata and then encrypting again with the right
offset.